refactor: [step2 피드백 반영] Session을 무료, 유료 강의로 분리
